relocate some definitions in prep for row-level ACL work  [ci skip]
We need to make use of more pieces w/o cyclic imports so move
some of the basic elements up into the model module...
