+++ ermrest/model/key.py
@@ -17,6 +17,7 @@
 from .. import exception
 from ..util import sql_identifier, sql_literal, constraint_exists
 from .misc import frozendict, AltDict, AclDict, keying, annotatable, commentable, hasacls, hasdynacls, enforce_63byte_id, truncated_identifier
+from .name import _keyref_join_str, _keyref_join_sql
 
 import json
 
@@ -377,27 +378,6 @@ def _guarded_add(s, new_fkr):
             return
     # otherwise this is a new leader
     s.add(new_fkr)
-
-def _keyref_join_str(self, refop, lname, rname):
-    if refop == '=@':
-        lcols = self._from_column_names()
-        rcols = self._to_column_names()
-    else:
-        lcols = self._to_column_names()
-        rcols = self._from_column_names()
-    return '%s:%s%s%s:%s' % (lname, ','.join(lcols), refop, rname, ','.join(rcols))
-
-def _keyref_join_sql(self, refop, lname, rname):
-    if refop == '=@':
-        lcols = self._from_column_names()
-        rcols = self._to_column_names()
-    else:
-        lcols = self._to_column_names()
-        rcols = self._from_column_names()
-    return ' AND '.join([
-        '%s.%s = %s.%s' % (lname, sql_identifier(lcols[i]), rname, sql_identifier(rcols[i]))
-        for i in range(len(lcols))
-    ])
 
 def _keyref_from_column_names(self):
     f_cnames = [ unicode(col.name) for col in self.foreign_key.columns ]
@@ -812,94 +792,3 @@ SELECT _ermrest.model_change_event();
     def has_right(self, aclname, roles=None):
         return _keyref_has_right(self, aclname, roles)
 
-class _Endpoint(object):
-
-    def __init__(self, table):
-        self.table = table
-                
-class MultiKeyReference (object):
-    """A disjunctive join condition collecting several links.
-
-       This abstraction simulates a left-to-right reference.
-    """
-
-    def __init__(self, links):
-        assert len(links) > 0
-        
-        self.ltable = None
-        self.rtable = None
-
-        for keyref, refop in links:
-            if refop == '=@':
-                ltable = keyref.foreign_key.table
-                rtable = keyref.unique.table
-            else:
-                ltable = keyref.unique.table
-                rtable = keyref.foreign_key.table
-                
-            assert self.ltable is None or self.ltable == ltable
-            assert self.rtable is None or self.rtable == rtable
-
-            self.ltable = ltable
-            self.rtable = rtable
-
-        self.links = links
-        self.foreign_key = _Endpoint(ltable)
-        self.unique = _Endpoint(rtable)
-
-    def _visible_links(self):
-        return [ l for l in self.links if l.has_right('enumerate') ]
-
-    def join_str(self, refop, lname='..', rname='.'):
-        """Build a simplified representation of the join condition."""
-        assert refop == '=@'
-        parts = []
-        for keyref, refop in self._visible_links():
-            parts.append(keyref.join_str(refop, lname, rname))
-        return '(%s)' % (' OR '.join(parts))
-                         
-    def join_sql(self, refop, lname, rname):
-        assert refop == '=@'
-        return ' OR '.join([
-            '(%s)' % keyref.join_sql(refop, lname, rname)
-            for keyref, refop in self._visible_links()
-        ])
-
-    def has_right(self, aclname, roles=None):
-        assert aclname == 'enumerate'
-        return self._visible_links()
-
-class ExplicitJoinReference (object):
-
-    def __init__(self, lcols, rcols):
-        assert len(lcols) == len(rcols)
-
-        self.lcols = lcols
-        self.rcols = rcols
-
-        ltable = lcols[0].table
-        rtable = rcols[0].table
-
-        self.foreign_key = _Endpoint(ltable)
-        self.unique = _Endpoint(rtable)
-
-    def _from_column_names(self):
-        return [ c.name for c in self.lcols ]
-
-    def _to_column_names(self):
-        return [ c.name for c in self.rcols ]
-        
-    def join_str(self, refop, lname='..', rname='.'):
-        assert refop == '=@'
-        return _keyref_join_str(self, refop, lname, rname)
-        
-    def join_sql(self, refop, lname, rname):
-        assert refop == '=@'
-        return _keyref_join_sql(self, refop, lname, rname)
-
-    def has_right(self, aclname, roles=None):
-        assert aclname == 'enumerate'
-        for c in self.lcols + self.rcols:
-            if not c.has_right(aclname, roles):
-                return False
-        return True
