+++ ermrest/model/makefile-vars
@@ -5,6 +5,7 @@ ERMREST_MODEL_PYTHON_FILES= \
 	introspect.py \
 	key.py \
 	misc.py \
+	name.py \
 	schema.py \
 	table.py \
 	type.py
