+++ ermrest/url/ast/data/__init__.py
@@ -26,7 +26,7 @@ import tempfile
 from ..api import Api
 from . import path
 from .predicate import predicatecls
-from ..name import Name
+from ....model.name import Name
 from .... import ermpath, exception
 from webauthn2.util import urlquote
