+++ ermrest/url/ast/makefile-vars
@@ -3,8 +3,7 @@ ERMREST_URL_AST_PYTHON_FILES= \
 	__init__.py \
 	api.py \
 	catalog.py \
-	model.py \
-	name.py
+	model.py
 
 ERMREST_URL_AST_PYTHON_FILES_INSTALL=$(ERMREST_URL_AST_PYTHON_FILES:%=$(PYLIBDIR)/ermrest/url/ast/%)
 
@@ -14,7 +13,8 @@ INSTALL_DIRS += $(PYLIBDIR)/ermrest/url/ast
 
 CLEAN_FILES += \
 	$(ERMREST_URL_AST_PYTHON_FILES:%=ermrest/url/ast/%c) \
-	$(ERMREST_URL_AST_PYTHON_FILES:%=ermrest/url/ast/%o)
+	$(ERMREST_URL_AST_PYTHON_FILES:%=ermrest/url/ast/%o) \
+	$(PYLIBDIR)/ermrest/url/ast/name.py
 
 EDIT_FILES += $(ERMREST_URL_AST_PYTHON_FILES:%=ermrest/url/ast/%) \
 	ermrest/url/ast/makefile-vars
